make JavaClobDao as an inner interface
