+++ pom.xml
@@ -256,7 +256,7 @@
         <dependency>
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-classic</artifactId>
-            <version>1.0.13</version>
+            <version>1.1.8</version>
             <scope>test</scope>
         </dependency>
